Feature/order detail

+++ src/main/java/com/onshop/shop/product/SellerProductsRequestDTO.java
@@ -30,7 +30,7 @@ public class SellerProductsRequestDTO {
 	
 	@Min(value = 0, message = "재고는 최소 0 이상이어야 합니다.")
 	@Max(value= 1000, message ="재고는 최대 1000이하이어야 합니다.")
-	private Integer stock; // 재고
+	private Long stock; // 재고
